Extract AbstractCommand
